+++ ChangeLog.md
@@ -7,7 +7,8 @@ The code was almost completely rewritten.
 * `--powerset-for-weak` for accepting SCC that is inherently weak in the input automaton, use only simple powerset construction in the deterministic part of the sDBA.
 * `--powerset-on-cut` starts the breakpoint/subset construction already on cut-transitions. For a marked transition s -a-> p build s -a-> (δ(s),δ_0(s),0) instead of s -a-> ({p},∅,0).
 * `--scc0` disables the SCC-aware optimization which is enabled by default
-* More transformation types can be chosen. The one will smallest result will be outputted.
+* Multiple transformation types (``--via-sba`, ``--via-tba`, ``--via-tgba`) can be specified. The one will smallest result will be outputted.
+* `--sd` option for semi-deterministic output (and not cut-deterministic)
 
 ### Changed
 * When a breakpoint is reached (R = B), track accepting transitions of the next level in B (it was set to ∅ before).